Crate proc_macro_api

Client-side Proc-Macro crate

We separate proc-macro expanding logic to an extern program to allow different implementations (e.g. wasm or dylib loading). And this crate is used to provide basic infrastructure for communication between two processes: Client (RA itself), Server (the external program)

Modules

bidirectional_protocol

Bidirectional protocol methods

legacy_protocol

The initial proc-macro-srv protocol, soon to be deprecated.

pool

A pool of proc-macro server processes

process

Handle process life-time and message passing for proc-macro client

transport

Contains construct for transport of messages.

version

The versions of the server protocol

Structs

MacroDylib

Represents a dynamically loaded library containing procedural macros.

ProcMacro

A handle to a specific proc-macro (a #[proc_macro] annotated function).

ProcMacroClient

A handle to proc-macro server process pool which load dylibs with macros (.so or .dll) and runs actual macro expansion functions.

ServerError

Represents errors encountered when communicating with the proc-macro server.

Enums

ProcMacroKind

Represents different kinds of procedural macros that can be expanded by the external server.

ProtocolFormat

Protocol format for communication between client and server.