proc_macro_api::legacy_protocol

Module msg

Defines messages for cross-process message passing based on ndjson wire protocol

Structs

ExpandMacro

Represents a macro expansion request sent from the client.

ExpandMacroData

Represents the input data required for expanding a macro.

ExpandMacroExtended

Represents an extended macro expansion response, including span data mappings.

ExpnGlobals

Represents global expansion settings, including span resolution.

FlatTree

PanicMessage

Represents an error message when a macro expansion results in a panic.

ServerConfig

Configuration settings for the proc-macro-srv.

TokenId

Legacy span type, only defined here as it is still used by the proc-macro server. While rust-analyzer doesn’t use this anymore at all, RustRover relies on the legacy type for proc-macro expansion.

Enums

Request

Represents requests sent from the client to the proc-macro-srv.

Response

Represents responses sent from the proc-macro-srv to the client.

SpanMode

Defines the mode used for handling span data.

Constants

CURRENT_API_VERSION

Current API version of the proc-macro protocol.

ENCODE_CLOSE_SPAN_VERSION

EXTENDED_LEAF_DATA

Whether literals encode their kind as an additional u32 field and idents their rawness as a u32 field.

HAS_GLOBAL_SPANS

NO_VERSION_CHECK_VERSION

RUST_ANALYZER_SPAN_SUPPORT

VERSION_CHECK_VERSION

Traits

Message

Functions

deserialize_span_data_index_map

serialize_span_data_index_map

Type Aliases

SpanDataIndexMap