Crate proc_macro_api

Client-side Proc-Macro crate

We separate proc-macro expanding logic to an extern program to allow different implementations (e.g. wasm or dylib loading). And this crate is used to provide basic infrastructure for communication between two processes: Client (RA itself), Server (the external program)

Modules

msg

Defines messages for cross-process message passing based on ndjson wire protocol

Structs

MacroDylib

MacroPanic

ProcMacro

A handle to a specific macro (a #[proc_macro] annotated function).

ProcMacroServer

A handle to an external process which load dylibs with macros (.so or .dll) and runs actual macro expansion functions.

RustCInfo

ServerError

Enums

ProcMacroKind

Functions

read_dylib_info

Read rustc dylib information

read_version

Check the version of rustc that was used to compile a proc macro crate’s