Crate proc_macro_api

Client-side Proc-Macro crate

We separate proc-macro expanding logic to an extern program to allow different implementations (e.g. wasm or dylib loading). And this crate is used to provide basic infrastructure for communication between two processes: Client (RA itself), Server (the external program)

Modules

legacy_protocol

Structs

MacroDylib

ProcMacro

A handle to a specific proc-macro (a #[proc_macro] annotated function).

ProcMacroClient

A handle to an external process which load dylibs with macros (.so or .dll) and runs actual macro expansion functions.

ServerError

Enums

ProcMacroKind