proc_macro_api::legacy_protocol

Module msg

Defines messages for cross-process message passing based on ndjson wire protocol

Structs

ExpandMacro

Represents a macro expansion request sent from the client.

ExpandMacroData

Represents the input data required for expanding a macro.

ExpandMacroExtended

Represents an extended macro expansion response, including span data mappings.

ExpnGlobals

Represents global expansion settings, including span resolution.

FlatTree

PanicMessage

Represents an error message when a macro expansion results in a panic.

ServerConfig

Configuration settings for the proc-macro-srv.

Enums

Request

Represents requests sent from the client to the proc-macro-srv.

Response

Represents responses sent from the proc-macro-srv to the client.

SpanMode

Defines the mode used for handling span data.

Traits

Message

SpanTransformer

Functions

deserialize_span_data_index_map

serialize_span_data_index_map

Type Aliases

SpanDataIndexMap