proc_macro_api::legacy_protocol
Module msg

Defines messages for cross-process message passing based on ndjson wire protocol

Structs

ExpandMacro

ExpandMacroData

ExpandMacroExtended

ExpnGlobals

FlatTree

PanicMessage

ServerConfig

TokenId

Legacy span type, only defined here as it is still used by the proc-macro server. While rust-analyzer doesn’t use this anymore at all, RustRover relies on the legacy type for proc-macro expansion.

Enums

Request

Response

SpanMode

Constants

CURRENT_API_VERSION

ENCODE_CLOSE_SPAN_VERSION

EXTENDED_LEAF_DATA

Whether literals encode their kind as an additional u32 field and idents their rawness as a u32 field

HAS_GLOBAL_SPANS

NO_VERSION_CHECK_VERSION

RUST_ANALYZER_SPAN_SUPPORT

VERSION_CHECK_VERSION

Traits

Message

Functions

deserialize_span_data_index_map

serialize_span_data_index_map

Type Aliases

SpanDataIndexMap